test: update scenarios for next branch
