+++ features/fixtures/mazerunner/cxx-scenarios-bugsnag/build.gradle
@@ -16,11 +16,11 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_6
-        targetCompatibility JavaVersion.VERSION_1_6
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
-        jvmTarget = "1.6"
+        jvmTarget = "1.8"
     }
     lintOptions {
         tasks.lint.enabled = false
