+++ features/fixtures/mazerunner/cxx-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/CXXSignalSmokeScenario.java
@@ -22,9 +22,7 @@ import java.util.Set;
 public class CXXSignalSmokeScenario extends Scenario {
 
     static {
-        System.loadLibrary("bugsnag-ndk");
-        System.loadLibrary("monochrome");
-        System.loadLibrary("entrypoint");
+        System.loadLibrary("cxx-scenarios");
     }
 
     public native int crash(int value);
