+++ features/fixtures/mazerunner/jvm-scenarios/src/main/resources/naughty_strings.txt
